Use StaticPicker component by default
