+++ spec/picker-virtual-scroll-spec.js
@@ -1,18 +1,8 @@
-// NOTE: this is a full copy of the picker-spec.js
-// The main change is in import - 'StaticPicker as Picker'
-// Other changes are marked with `// StaticPicker change` comments.
 import { mount } from '@vue/test-utils'
 
 import data from '../data/all.json'
 import { EmojiIndex } from '../src/utils/emoji-data'
-import {
-  Anchors,
-  Search,
-  StaticPicker as Picker,
-  Category,
-  Preview,
-  Emoji,
-} from '../src/components'
+import { Anchors, Picker, Category, Preview, Emoji } from '../src/components'
 
 describe('Picker', () => {
   let index = new EmojiIndex(data)
@@ -44,12 +34,11 @@ describe('Picker', () => {
     expect(picker.html()).toContain('woman-gesturing-ok"')
   })
 
-  it('renders 10 categories', () => {
+  it('renders 8 categories', () => {
     // Due to the virtual scroller, not all the categories
     // are rendered at once
     let categories = picker.findAll(Category)
-    // StaticPicker change (8 -> 10)
-    expect(categories.length).toBe(11)
+    expect(categories.length).toBe(8)
     // Hidden category with search results
     expect(categories.at(0).vm.name).toBe('Search')
     expect(categories.at(1).vm.name).toBe('Recent')
@@ -59,39 +48,6 @@ describe('Picker', () => {
     expect(categories.at(5).vm.name).toBe('Food & Drink')
     expect(categories.at(6).vm.name).toBe('Activities')
     expect(categories.at(7).vm.name).toBe('Travel & Places')
-    expect(categories.at(8).vm.name).toBe('Objects')
-    // StaticPicker change (Symbols and Flags)
-    expect(categories.at(9).vm.name).toBe('Symbols')
-    expect(categories.at(10).vm.name).toBe('Flags')
-  })
-
-  it('no error when clicking anchor when search is active', (done) => {
-    let search = picker.find(Search)
-    let input = search.find('input')
-    input.element.value = '+1'
-    input.trigger('input')
-
-    picker.vm.$nextTick(() => {
-      let categories = picker.findAll(Category)
-      let searchCategory = categories.at(0)
-      expect(searchCategory.vm.id).toBe('search')
-
-      let anchors = picker.find(Anchors)
-      let anchorsCategories = anchors.findAll('span.emoji-mart-anchor')
-      let symbols = anchorsCategories.at(8)
-      expect(symbols.element.attributes['data-title'].value).toBe('Symbols')
-      symbols.trigger('click')
-
-      picker.vm.$nextTick(() => {
-        let events = anchors.emitted().click
-        let category = events[0][0]
-        expect(category.id).toBe('symbols')
-        // Category didn't not change (search as active)
-        expect(anchors.vm.activeCategory.id).toBe('recent')
-
-        done()
-      })
-    })
   })
 })
 
@@ -222,8 +178,7 @@ describe('anchors', () => {
     expect(category.id).toBe('symbols')
     expect(category.name).toBe('Symbols')
 
-    // StaticPicker change - the check below fails (although works in demo app)
-    // expect(anchors.vm.activeCategory.id).toBe('symbols')
+    expect(anchors.vm.activeCategory.id).toBe('symbols')
   })
 })
