+++ src/components/index.js
@@ -5,4 +5,3 @@ export { default as Search } from './search'
 export { default as Skins } from './skins'
 export { default as Emoji } from './Emoji'
 export { default as Picker } from './Picker'
-export { default as StaticPicker } from './StaticPicker'
